Fix sandbox filtering on projects page
It seems that the database schema was updated: projects now have
a database name (string) which is either "" (meaning the main OSM
database) or else the name of a sandbox. This commit updates the API and
frontend so that filtering the projects page by type (OSM or Sandbox)
works again.