Fix sandbox authentication
Previously, the sandbox-iD editor authenticated to the sandbox server
using the popup flow, which used a cookie set on the sandbox domain
to prompt the user to authorize the client and then retrieve a session
token.

This authentication pathway is no longer supported by the OSM Rails app
for security reasons [1].

To work around this, the sandbox Dashboard API has been enhanced to support
server-to-server token exchange. When a user authenticates to the Dashboard
with their OSM credentials, the Dashboard retrieves an OAuth token for the
given sandbox using the password grant flow. The client can then request
this token using the session ID that the Dashboard assigns them.

To simplify the sandbox-iD editor, communication with the Dashboard is done
by the Tasking Manager frontend. The editor doesn't know anything about the
complex authentication system; it just recieves an API URL and a valid token
for that API when it is initialized.

This will make it easier to integrate other editors (like Rapid) in the future.

[1]: https://github.com/openstreetmap/openstreetmap-website/commit/2ff4d6a4